+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * [artifact-manager] IAC-790 / Update usage of deprecated policy APIs for vROPs 8.12.x.
 * [archetypes] IAC-797 / Improved content.yaml definition for "ts-vra-ng" and "vra-ng"
 
+### Bugs
+* [o11n-plugin-ssh] session esxiCode returns type void() instead of number
+
 ## v2.35.1 - 29 Sep 2023
 
 ### Fixes