+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Enhancements
 * [artifact-manager] IAC-786 / Set Aria Operations Default Policy vROPs 8.12.x.
 * [artifact-manager] IAC-790 / Update usage of deprecated policy APIs for vROPs 8.12.x.
+* [archetypes] IAC-797 / Improved content.yaml definition for "ts-vra-ng" and "vra-ng"
 
 ### Bugs
 * [o11n-plugin-ssh] session esxiCode returns type void() instead of number